CRUD User + Authentication

+++ ShoppingCard/Models/CartItemModel.cs
@@ -3,7 +3,8 @@ namespace ShoppingCard.Models
 {
     public class CartItemModel
     {
-        public long ProductId { get; set; }
+        public int ProductId { get; set; } // Nhoi quay lai check tu bai 42
+
         public string ProductName { get; set; }
         public int Quantity { get; set; }
         public decimal Price { get; set; }
